Speed up drawing: track peak memory usage
Track peak memory usage in the entire data set (and, separately, just
for the visible portion, to avoid scale jumps if the process grows while
the graph is paused).

This way we don't have to compute the maximum value on every frame.

The code assumes that VIRT is always greater than or equal to RSS.